pyproject.toml
[project]
name = "mee6-discord-bot"
version = "1.1.0"
description = "Многофункциональный Discord бот"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "aiohttp>=3.9.1",
    "python-dateutil>=2.8.2",
    "asyncio>=3.4.3",
    "SQLAlchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.1",
    "sentry-sdk>=1.40.0",
    "prometheus-client>=0.19.0",
    "PyNaCl>=1.5.0",
    "aiosqlite>=0.19.0",
    "aiocache>=0.12.2",
    "motor>=3.3.2",
    "ujson>=5.8.0",
    "backoff>=2.2.1",
    "pycryptodome>=3.19.1",
    "cryptography>=42.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
